Change behavior of __is_in() function
The current behavior of the `__is_in()` function is
to simply `return 0` when a match is found, or
`return 1`, when there's none. If limited to only
it, the library will need to parse the string
lists twice: firstly for the validity of options,
and secondly for the argument type (`reqarg` or
`noarg`).

Change the behavior of `__is_in()` to, on success,
print to stdout the argument type of the option,
so that, if one parsing, both the validity and the
type argument of an option are reached.