Change `assertTrue()` call in tests
The tests that use the `asserTrue()` are not
testing what is desired, because of the
expresssion:

`test $(command)`

What this expression evaluates is if the `stdout`
of `command` is not empty, that is,
`[ -n "$(command)" ]`. What should be tested is
whether or not these functions return 0.

Change the `assertTrue()` calls in the tests by
moving up one line the function that was being
called inside a subshell, and passing to the
former the return value of the latter with "${?}".