Add check for argless options in `__lopt_parse()`
One of the possible errors in which an argless
option may incur is to passed in the `key=value`
format. To avoid any misbehavior if such cases
were not handled, they are treated as invalid
options.

To achieve such an evaluation of invalidity, the
variables `${__lopt_parse_opt}` and
`${_lopt_parse_arg}` must be compared.

The former is the first argument of the function
after the following string deletions:

`__lopt_parse_opt="${1#--}"`
`__lopt_parse_opt="${__lopt_parse_opt%%=*}"`

The latter:
`__lopt_parse_opt="${1#*=}"

Which means that, when the option not passed in
the `key=value` format, these two variables will
be the same, as long as the first have a "--"
prefixed to it. Thus, by checking for their
inequality, it can be identified when an argless
option was passed in the aforementioned format.