Readd `shfmt` test
The versions `shfmt` in the snap store and on the
local machine where the library is written were
not same, respectively, `3.2.1` and `3.5.1`.
Perhaps that was the reason that, on the local
machine the `shfmt` command would succeed, while
on the Travis-Ci's vm it would fail.

Install `shfmt` from the Github's instead of the
snap package.